+++ src/app/(causw)/home/ClientGraduatePage.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-import { Banner } from '@/entities/home';
+import { Banner } from '@/fsd_entities/home';
 
 import { HomeRscService } from '@/shared';
