--- src/entities/home/index.ts
@@ -1,4 +1,0 @@
-export * from './Banner';
-export * from './card/CardBox';
-export * from './card/HomeCard';
-export * from './card/LinkCard';
