+++ src/fsd_entities/home/ui/Banner/Banner.tsx
