+++ src/fsd_entities/home/ui/Banner/index.ts
@@ -1,0 +1,1 @@
+export { Banner } from './Banner';
