+++ src/fsd_entities/home/ui/CardBox/CardBox.tsx
