+++ src/fsd_entities/home/ui/CardBox/index.ts
@@ -1,0 +1,1 @@
+export { CardBox } from './CardBox';
