+++ src/fsd_entities/home/ui/HomeCard/index.ts
@@ -1,0 +1,1 @@
+export { HomeCard, type HomeCardProps } from './HomeCard';
