+++ src/fsd_entities/home/ui/LinkCard/index.ts
@@ -1,0 +1,1 @@
+export { LinkCard } from './LinkCard';
